controller/pyproject.toml
[project]
name = "controller"
version = "0.1.0"
description = "Controller for Problemologist AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "fastapi>=0.128.0",
    "langchain-core>=0.3.35",
    "langchain-openai>=0.3.6",
    "langgraph>=0.2.72",
    "langgraph-checkpoint-sqlite>=2.0.3",
    "langgraph-checkpoint-postgres>=2.0.12",
    "psycopg[binary]>=3.2.4",
    "langfuse>=2.59.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "deepagents>=0.3.10",
    "httpx>=0.27.0",
    "temporalio>=1.11.0",
    "asyncpg>=0.30.0",
    "boto3>=1.36.0",
    "typer>=0.15.1",
    "streamlit>=1.42.0",
    "streamlit-autorefresh>=1.0.1",
    "stpyvista>=0.0.8",
    "shared",
]

[tool.uv.sources]
shared = { workspace = true }

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
